requirements.txt
fastapi
uvicorn
pyyaml
pycryptodome
requests
python-multipart
